+++ NEWS.md
@@ -3,10 +3,50 @@ This file documents the development of the package as well as open issues or poi
 
 
 
+
+### Version in 1.2.2 (fifth release)
+Date: 2019.09.18
+
+Changes:
+
+*    Error in rankByGroup and rankBySlidingWindow when covariate variable name was used
+*    warning added to rankBySlidingWindow in case, age and group do not correspond
+*    rankBySlidingWindow accepts age variable in addition to age variable name
+*    plotPercentiles now allows plotting both degrees of binary covariate
+*    code cleaning in plotPercentiles
+*    bestModel automatically does plotPercentiles if parameter plot set to TRUE
+*    Citation file added
+
+
+
+
+
+### Version in 1.2.1
+Date: 2019.08.01
+
+Changes:
+
+*    prepareDate issues warnings if age and group values do not relate
+*    modified message in bestModel function if R2 is not reached
+*    if R2 and terms are not specified in bestModel function, fall back to model 5
+     in case R2 does not reach .99 in most complex model
+*    Additional message in bestModel in case of high number of terms     
+*    Code simplification in computePowers by using 'poly' function
+*    Performance improvement and code cleaning in checkConsistency
+*    Multiple R2 output added to computePowers
+*    new option to add index labels to data points in plotSubset
+*    cnorm.cv now respects sliding window ranking
+*    silent parameter added to prepareData and computePowers
+*    rankBySlidingWindo and rankByGroup add width parameter to data preparation
+*    pCutoff in cnorm.cv now adjusts for sample size
+
+
+
 ### Version in 1.2.0 (fourth release)
 Date: 2019.07.26
 
 Changes:
+
 *   Preparing for the inclusion of a binary covariate. The package is been
     rewritten by larger parts
     NOTE: The inclusion of a covariate is currently still experimental and not
@@ -24,6 +64,7 @@ Changes:
 Date: 2019.07.07
 
 Changes:
+
 *   weighting added to bestModel-function
 *   cast to data.frame in prepareData method to prevent SPSS import failing
 *   Ordering of raw table when using descending values
@@ -38,6 +79,7 @@ Changes:
 Date: 2019.03.15 (mainly testing and cleaning minor errors)
 
 Changes:
+
 *   fixing exceptions
 *   group and age can now be deactivated, resulting in conventional norming procedure, based in ranking + regression over powers of L
 
@@ -47,6 +89,7 @@ Changes:
 Date: 2019.02.28
 
 Changes:
+
 *   bestModel function now accepts a formula as a predictors object
 *   plotPercentile now accepts descending ranking
 *   rawTable and normTable adapted for descending values
@@ -62,6 +105,7 @@ Changes:
 Date: 2019.02.07, third release on CRAN
 
 Changes:
+
 *   Improvement to cv function in GUI and in package
 
 
@@ -70,6 +114,7 @@ Changes:
 Date: 2019.02.06, third release on CRAN
 
 Changes:
+
 *   Cross Validation added to shiny GUI
 *   cnorm.cv documentation improved
 *   added information to BestModel output
@@ -81,6 +126,7 @@ Changes:
 Date: 2018.12.18
 
 Changes:
+
 *   scale parameter added to prepareData function
 *   fix for plotNorm by group with missing values
 
@@ -90,6 +136,7 @@ Changes:
 Date: 2018.12.09
 
 Changes:
+
 *   rmarkdown moved from imports to suggests
 *   cnorm.cv info added to README
 
@@ -99,6 +146,7 @@ Changes:
 Date: 2018.12.08
 
 Changes:
+
 *   deleted code in vignette needing to much build time
 *   removed UTF-8 attributes from ppvt dataset and cleared all datasets from non ASCII signs
 *   deleted code in vignette needing to much build time
@@ -111,6 +159,7 @@ Changes:
 Date: 2018.12.01
 
 Changes:
+
 *    Parameters added to cv.norm: Significance level for stratification process
 *    Additional plot in cv.norm: delta R2 in norm score validation
 *    Example in readme improved